+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## [0.1.2] - 2026-02-18
+
+### Summary
+
+refactor(docs): configuration management system
+
+### Docs
+
+- docs: update README
+
+### Build
+
+- update pyproject.toml
+
+### Config
+
+- config: update goal.yaml
+
+### Other
+
+- update .idea/inspectionProfiles/Project_Default.xml
+- update .idea/inspectionProfiles/profiles_settings.xml
+- update .idea/marksync.iml
+- update .idea/misc.xml
+- update .idea/modules.xml
+- update .idea/vcs.xml
+- update engine.py
+- update marksync/dsl/__init__.py
+- update marksync/dsl/executor.py
+- update marksync/dsl/parser.py
+
+
 ## [0.1.1] - 2026-02-18
 
 ### Summary